requirements.txt
pygame>=2.5.0
python-chess>=1.999
cairosvg>=2.7.0
pillow>=10.0.0